adds exec and logs to join e2e test

+++ test/e2e/suite/nodeadm_test.go
@@ -8,6 +8,7 @@ import (
 	"flag"
 	"fmt"
 	"os"
+	"strings"
 	"testing"
 	"time"
 
@@ -26,6 +27,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	clientgo "k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	"sigs.k8s.io/yaml"
 
@@ -79,17 +81,18 @@ func TestE2E(t *testing.T) {
 }
 
 type peeredVPCTest struct {
-	aws         awsconfig.Config // TODO: move everything to aws sdk v2
-	awsSession  *session.Session
-	eksClient   *eks.EKS
-	ec2Client   *ec2v1.EC2
-	ec2ClientV2 *ec2v2.Client
-	ssmClient   *ssmv1.SSM
-	ssmClientV2 *ssmv2.Client
-	cfnClient   *cloudformation.CloudFormation
-	k8sClient   *clientgo.Clientset
-	s3Client    *s3v1.S3
-	iamClient   *iam.IAM
+	aws             awsconfig.Config // TODO: move everything to aws sdk v2
+	awsSession      *session.Session
+	eksClient       *eks.EKS
+	ec2Client       *ec2v1.EC2
+	ec2ClientV2     *ec2v2.Client
+	ssmClient       *ssmv1.SSM
+	ssmClientV2     *ssmv2.Client
+	cfnClient       *cloudformation.CloudFormation
+	k8sClient       *clientgo.Clientset
+	k8sClientConfig *rest.Config
+	s3Client        *s3v1.S3
+	iamClient       *iam.IAM
 
 	logger     logr.Logger
 	logsBucket string
@@ -259,6 +262,7 @@ var _ = Describe("Hybrid Nodes", func() {
 							Expect(err).NotTo(HaveOccurred(), "EC2 Instance should have been created successfully")
 
 							joinNodeTest := joinNodeTest{
+								clientConfig:  test.k8sClientConfig,
 								k8s:           test.k8sClient,
 								nodeIPAddress: instance.IP,
 								logger:        test.logger,
@@ -336,6 +340,7 @@ var _ = Describe("Hybrid Nodes", func() {
 							Expect(err).NotTo(HaveOccurred(), "EC2 Instance should have been created successfully")
 
 							joinNodeTest := joinNodeTest{
+								clientConfig:  test.k8sClientConfig,
 								k8s:           test.k8sClient,
 								nodeIPAddress: instance.IP,
 								logger:        test.logger,
@@ -492,6 +497,7 @@ func (c createNodeTest) Run(ctx context.Context) (ec2.Instance, error) {
 }
 
 type joinNodeTest struct {
+	clientConfig  *rest.Config
 	k8s           *clientgo.Clientset
 	nodeIPAddress string
 	logger        logr.Logger
@@ -520,6 +526,29 @@ func (t joinNodeTest) Run(ctx context.Context) error {
 		return err
 	}
 	t.logger.Info(fmt.Sprintf("Pod %s created and running on node %s", podName, nodeName))
+
+	t.logger.Info("Exec-ing nginx -version", "pod", podName)
+	stdout, stderr, err := kubernetes.ExecPod(ctx, t.clientConfig, t.k8s, podName, podNamespace, "/sbin/nginx", "-version")
+	if err != nil {
+		return err
+	}
+	if !strings.Contains(stdout, "nginx") {
+		return fmt.Errorf("pod exec stdout does not contain expected value %s: %s", stdout, "nginx")
+	}
+	if stderr != "" {
+		return fmt.Errorf("pod exec stderr should be empty %s", stderr)
+	}
+	t.logger.Info("Successfully exec'd nginx -version", "pod", podName)
+
+	t.logger.Info("Checking logs for nginx output", "pod", podName)
+	logs, err := kubernetes.GetPodLogs(ctx, t.k8s, podName, podNamespace)
+	if err != nil {
+		return err
+	}
+	if !strings.Contains(logs, "nginx") {
+		return fmt.Errorf("pod log does not contain expected value %s: %s", logs, "nginx")
+	}
+	t.logger.Info("Successfully validated log output", "pod", podName)
 
 	t.logger.Info("Deleting test pod", "pod", podName)
 	if err = kubernetes.DeletePod(ctx, t.k8s, podName, podNamespace); err != nil {
@@ -647,6 +676,7 @@ func buildPeeredVPCTestForSuite(ctx context.Context, suite *suiteConfiguration) 
 	if err != nil {
 		return nil, err
 	}
+	test.k8sClientConfig = clientConfig
 	test.k8sClient, err = clientgo.NewForConfig(clientConfig)
 	if err != nil {
 		return nil, err
